Add files via upload
WEPP simulation was conducted in for 122 scenarios of representative hillslopes for the Lower Puyallup River Watershed. This code extracts the water balance and other outputs for each of the scenarios and analyse the change in water balance in various scenarios. It also compute the HSI values for each hillslope and map it with runoff. 

This code also analyze the difference in runoff generation in No treatment Vs Pavement Vs Rain Garden scenarios.